+++ crates/mzoon/src/command/build.rs
@@ -1,8 +1,8 @@
 use anyhow::Error;
 use fehler::throws;
 use crate::config::Config;
-use crate::frontend::build_frontend;
-use crate::backend::build_backend;
+use crate::build_frontend::build_frontend;
+use crate::build_backend::build_backend;
 use crate::set_env_vars::set_env_vars;
 
 #[throws]
